Add environment builds
