+++ .gitignore
@@ -1,8 +1,8 @@
 # dependencies
 node_modules
 
-# production
-lib
+# bundles
+cjs
 
 # misc
 .DS_Store
