+++ package.json
@@ -2,20 +2,20 @@
   "name": "react-native-use-styles",
   "version": "1.2.3",
   "description": "React Native useStyles",
-  "main": "lib/index.js",
+  "main": "index.js",
   "module": "src/index.js",
   "types": "src/index.d.ts",
   "files": [
     "src",
-    "lib",
+    "cjs",
     "README.md"
   ],
   "scripts": {
     "test": "jest --verbose",
     "test:watch": "jest --verbose --watch",
     "test:coverage": "jest --collect-coverage",
-    "build": "rollup -c",
-    "prepare": "rollup -c",
+    "build": "cross-env NODE_ENV=production rollup -c && cross-env NODE_ENV=development rollup -c",
+    "prepare": "cross-env NODE_ENV=production rollup -c && cross-env NODE_ENV=development rollup -c",
     "lint": "eslint *.js **/*.js",
     "check:types": "tsc"
   },
@@ -45,6 +45,7 @@
     "@rollup/plugin-replace": "^2.3.3",
     "@testing-library/react-native": "^7.0.2",
     "babel-eslint": "^10.1.0",
+    "cross-env": "^7.0.2",
     "eslint": "^6.6.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-import": "^2.18.2",
