+++ rollup.config.js
@@ -1,14 +1,13 @@
 import replace from '@rollup/plugin-replace';
 import resolve from '@rollup/plugin-node-resolve';
-import pkg from './package.json';
 
 export default {
   input: 'src/index.js',
   external: ['react', 'react-native'],
-  output: [{ file: pkg.main, format: 'cjs' }],
+  output: [{ file: `./cjs/${process.env.NODE_ENV}.js`, format: 'cjs' }],
   plugins: [
     resolve(),
-    replace({
+    process.env.NODE_ENV === 'production' && replace({
       'process.env.NODE_ENV': JSON.stringify('production')
     })
   ]
