+++ src/utils/index.js
@@ -5,7 +5,6 @@ import {
   COMPTUED_PREFIX,
   NAMESPACE_REGEX
 } from "../constants";
-import { separator } from "../core/transformer";
 
 export const isNamespace = style => style.startsWith(NAMESPACE_PREFIX);
 
@@ -15,8 +14,6 @@ export const getKey = style => style.substring(1);
 
 export const getNamespace = style =>
   style.match(NAMESPACE_REGEX)[0].substring(1);
-
-export const hasPath = style => style.indexOf(separator) !== -1;
 
 export const hasClassName = style => style.indexOf(CLASS_PREFIX) !== -1;
